pyproject.toml
[tool.poetry]
name = "tcc"
version = "0.1.0"
description = "試験項目書生成ツール"
authors = ["Toshihiro Yoshida <good.che.ex.mu2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
click = "^8.1.3"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
